+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - 调整渲染块全屏模式时的样式 | Adjust the style when rendering the block in full screen mode.
 - 为块全屏显示模式添加滚动条 | Add scroll bar to block full screen display mode.
 - 调整部分按钮图标 | Adjust some button icons.
+- 实现 jupyter 导入的 markdown 单元格解析 | Implement jupyter imported markdown cell parsing.
 
 ## v1.4.0/2022-11-06